0x0A: task 3

+++ 0x0A-argc_argv/3-mul.c
@@ -10,7 +10,7 @@
 
 int main(int argc, char *argv[])
 {
-	int n1, n2;
+	long int n1, n2;
 
 	if (argc == 1)
 		printf("Error\n");
@@ -19,7 +19,7 @@ int main(int argc, char *argv[])
 		n1 = atoi(argv[1]);
 		n2 = atoi(argv[2]);
 
-		printf("%d\n", n1 * n2);
+		printf("%ld\n", n1 * n2);
 	}
 	return (0);
 }
